feat: auth_handlerの切り分けとerrorResponseをmodel層で定義
