+++ model/error.go
@@ -1,0 +1,9 @@
+package model
+
+type ErrorResponse struct {
+	Error string `json:"error"`
+}
+
+func NewErrorResponse(message string) ErrorResponse {
+	return ErrorResponse{Error: message}
+}
